pyproject.toml
[project]
name = "orm-tips-pysul-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-ckeditor>=6.7.3",
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "django-silk>=5.4.3",
    "ipython>=9.7.0",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "pre-commit>=4.4.0",
    "ruff>=0.14.5",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
